requirements.txt
celery==5.3.6
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.14.0
psycopg2==2.9.9
pillow==10.2.0
gunicorn==20.1.0
django-environ==0.8.1
psycopg2-binary==2.9.2

